requirements.txt
# Core dependencies
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0

# Google Agent SDK
google-generativeai>=0.3.0

# LLM utilities
langchain>=0.0.300
numpy>=1.24.0
pandas>=2.0.0

# Git analysis
GitPython>=3.1.30

# Vector database
chromadb>=0.4.15

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Dev tools
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
